finshed clubs tab with some bugs and pending scroll section and pending googlr maps api
