+++ src/App.css
@@ -1,9 +1,11 @@
 /* @import url('https://fonts.googleapis.com/css2?family=Quicksand:wght@500&display=swap'); */
-@import url('https://fonts.googleapis.com/css2?family=Eczar&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Manrope&display=swap');
+/* @import url('https://fonts.googleapis.com/css2?family=Eczar&display=swap'); */
 
 *{
   /* font-family: 'Quicksand', sans-serif; */
-  font-family: 'Eczar', serif;
+  font-family: 'Manrope', sans-serif;
+  /* font-family: 'Eczar', serif; */
 }
 
 .splash{
